sli: move small read duration metrics to store/tikv (#25343)
